backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.2.1
pydantic>=2.0
pydantic-settings>=2.0
pytest>=7.0
httpx>=0.25.0
